kafka-payment-labs/day-2/lab-03-fastapi-integration/solution/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
confluent-kafka[avro]==2.3.0
pydantic==2.5.0
python-dotenv==1.0.0